+++ packages/@expo/config-plugins/src/ios/IosConfig.types.ts
@@ -26,11 +26,6 @@ export type InfoPlist = Record<string, JSONValue | undefined> & {
   CFBundleDevelopmentRegion?: string;
   ITSAppUsesNonExemptEncryption?: boolean;
   LSApplicationQueriesSchemes?: string[];
-  FacebookAppID?: string;
-  FacebookDisplayName?: string;
-  FacebookAutoInitEnabled?: boolean;
-  FacebookAutoLogAppEventsEnabled?: boolean;
-  FacebookAdvertiserIDCollectionEnabled?: boolean;
   UIBackgroundModes?: string[];
   UISupportedInterfaceOrientations?: InterfaceOrientation[];
   GMSApiKey?: string;