+++ packages/@expo/prebuild-config/src/plugins/withDefaultPlugins.ts
@@ -20,7 +20,6 @@ import withAppleAuthentication from './unversioned/expo-apple-authentication';
 import withBranch from './unversioned/expo-branch/expo-branch';
 import withContacts from './unversioned/expo-contacts';
 import withDocumentPicker from './unversioned/expo-document-picker';
-import withFacebook from './unversioned/expo-facebook/expo-facebook';
 import withNavigationBar from './unversioned/expo-navigation-bar/expo-navigation-bar';
 import withNotifications from './unversioned/expo-notifications/expo-notifications';
 import withSplashScreen from './unversioned/expo-splash-screen/expo-splash-screen';
@@ -32,7 +31,7 @@ const debug = Debug('expo:prebuild-config');
 
 /**
  * Config plugin to apply all of the custom Expo iOS config plugins we support by default.
- * TODO: In the future most of this should go into versioned packages like expo-facebook, expo-updates, etc...
+ * TODO: In the future most of this should go into versioned packages like expo-updates, etc...
  */
 export const withIosExpoPlugins: ConfigPlugin<{
   bundleIdentifier: string;
@@ -69,7 +68,7 @@ export const withIosExpoPlugins: ConfigPlugin<{
 
 /**
  * Config plugin to apply all of the custom Expo Android config plugins we support by default.
- * TODO: In the future most of this should go into versioned packages like expo-facebook, expo-updates, etc...
+ * TODO: In the future most of this should go into versioned packages like expo-updates, etc...
  */
 export const withAndroidExpoPlugins: ConfigPlugin<{
   package: string;
@@ -133,7 +132,6 @@ const versionedExpoSDKPackages: string[] = [
   'expo-branch',
   'expo-navigation-bar',
   'expo-document-picker',
-  'expo-facebook',
   'expo-splash-screen',
   'expo-system-ui',
 ];
@@ -151,7 +149,6 @@ export const withVersionedExpoSDKPlugins: ConfigPlugin<{ expoUsername: string | 
     [withUpdates, { expoUsername }],
     withBranch,
     withDocumentPicker,
-    withFacebook,
     // System UI must come before splash screen as they overlap
     // and splash screen will warn about conflicting rules.
     withSystemUI,
@@ -184,7 +181,6 @@ const legacyExpoPlugins = [
   'expo-dev-client',
   'expo-image-picker',
   'expo-file-system',
-  'expo-ads-facebook',
   'expo-location',
   'expo-media-library',
   'expo-screen-orientation',